requirements.txt
Django==4.2.25
whitenoise==6.11.0
python-dotenv==1.1.1
supabase
supabase-auth
gunicorn==23.0.0
